% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailedPCA.R
\name{detailedPCA}
\alias{detailedPCA}
\title{Detailed PCA}
\usage{
detailedPCA(X, center = TRUE, scale = TRUE)
}
\arguments{
\item{X}{A complex matrix (or data frame) that serves as the data for principal
components analysis, featuring variables arranged in columns and observations in rows.}

\item{center}{A logical value (default = TRUE) indicating whether the variables should
be shifted to be zero centered.}

\item{scale}{A logical value (default = TRUE) indicating whether the variables should
be scaled to have unit variance before the analysis takes place.}
}
\value{
A list containing the following components: \describe{
\item{X}{The value of the rotated data (the centred (and scaled if requested) data
multiplied by the rotation matrix).}
\item{sdev}{The standard deviations of the principal components, calculated as the
square roots of the eigenvalues of the covariance/correlation matrix. The calculation
is actually done with the singular values of the data matrix.}
\item{eigval}{The eigenvalues of the covariance/correlation matrix, representing the
variance of the principal components. The calculation is actually done with the
singular values of the data matrix.}
\item{pctvar}{The percentage of variance explained by each principal component.}
\item{rotation}{The matrix of variable loadings (i.e., a matrix whose columns contain
the eigenvectors).}
\item{indexload}{The index of loadings for the variables.}
\item{correlation}{The correlations of the variables with each principal component.}
\item{Psi}{The Psi index, which depends on the magnitude of the eigenvalues taken from
the correlation matrix of the data set.
\deqn{\Psi = \sum(\lambda_i-1)^2,}
where \eqn{\lambda} is the eigenvalue.}
\item{Phi}{The Phi statistic, which measures the average level of correlation among
the variables.
\deqn{\Phi = \sqrt{\frac{\sum\lambda_i^2-p}{p(p-1)}},}
where \eqn{\lambda} is the eigenvalue, and \eqn{p} is the number of variables.}
\item{center}{The centering used, or \code{FALSE}.}
\item{scale}{The scaling used, or \code{FALSE}.}
}
}
\description{
Detailed statistics on principal components analysis (PCA).
}
\references{
\itemize{
\item Gleason, T. C. and Staelin, R. (1975).
A Proposal for Handling Missing Data.
\emph{Psychometrika}, 40(2), 229--252.
\item Mardia, K. V., Kent, J. T. and Bibby, J. M. (1979).
Multivariate Analysis. London, UK: Academic Press.
\item Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988).
The New S Language. Wadsworth & Brooks/Cole.
\emph{Computer Science Series, Pacific Grove, CA}.
\item Vieira, Vasco M. N. C. S. (2012).
Permutation Tests to Estimate Significances on Principal Components Analysis.
\emph{Computational Ecology and Software}, 2(2), 103--123.
\item Venables, W. N. and Ripley, B. D. (2013).
Modern Applied Statistics with S-PLUS.
Springer Science & Business Media.
}
}
